Initial clean up of localization system. Seperated particle filter into seperate class. Added localization param file.
