+++ CMakeLists.txt
@@ -232,3 +232,4 @@ add_subdirectory(src/movement)
 add_subdirectory(src/sensors)
 add_subdirectory(src/controllers)
 add_subdirectory(src/vision)
+add_subdirectory(src/localization)
